Refactor AGENT to CASEWORKER for compliance
