+++ .devcontainer/devcontainer.env.sample
@@ -54,8 +54,7 @@ KEYCLOAK_SECRET=
 
 # Gov.uk Notify -------------------------
 NOTIFY_KEY=
-NOTIFY_TEMPLATE=
-AGENT_TEMPLATE_ID=
-AGENT_EMAIL=
+CASEWORKER_TEMPLATE_ID=
+CASEWORKER_EMAIL=
 CUSTOMER_TEMPLATE_ID=
 # end Gov.uk Notify ---------------------
