+++ config.js
@@ -7,8 +7,8 @@ module.exports = {
   dateTimeFormat: 'DD MMM YYYY HH:mm:ss',
   email: {
     notifyApiKey: process.env.NOTIFY_KEY,
-    agentTemplateId: process.env.AGENT_TEMPLATE_ID,
-    agentEmail: process.env.AGENT_EMAIL,
+    caseworkerTemplateId: process.env.CASEWORKER_TEMPLATE_ID,
+    caseworkerEmail: process.env.CASEWORKER_EMAIL,
     customerTemplateId: process.env.CUSTOMER_TEMPLATE_ID
   },
   survey: {
